feat/#2 :recycle: Divide interface & class implementation
